+++ src/examples/worker/app.ts
@@ -1,6 +1,6 @@
 import { Component, stateOf, interfaceOf } from '../../core'
 import { StyleGroup } from '../../utils/style'
-import { ViewInterface } from '../../interfaces/view'
+import { View } from '../../interfaces/view'
 import h from 'snabbdom/h'
 
 let name = 'Main'
@@ -9,7 +9,7 @@ let components = {
   counter: require('./counter').default,
 }
 
-let view: ViewInterface = (ctx, s) => {
+let view: View = (ctx, s) => {
   let style = ctx.groups.style
   return h('div', {
     key: name,