+++ src/examples/worker/counter.ts
@@ -1,7 +1,7 @@
 import { Context, Component, ev } from '../../core'
 import { StyleGroup } from '../../utils/style'
 
-import { ViewInterface } from '../../interfaces/view'
+import { View } from '../../interfaces/view'
 import h from 'snabbdom/h'
 
 let name = 'Counter'
@@ -27,7 +27,7 @@ let inputs = (ctx: Context) => ({
   inc: () => actions.Inc(),
 })
 
-let view: ViewInterface = (ctx, s) => {
+let view: View = (ctx, s) => {
   let style = ctx.groups['style']
   return h('div', {
     key: name,